+++ InnerEye/ML/metrics_dict.py
@@ -21,13 +21,14 @@ from InnerEye.Common.common_util import check_properties_are_not_none
 from InnerEye.Common.metrics_constants import INTERNAL_TO_LOGGING_COLUMN_NAMES, LoggingColumns, MetricType, \
     MetricTypeOrStr, SEQUENCE_POSITION_HUE_NAME_PREFIX
 from InnerEye.ML.common import ModelExecutionMode
-from InnerEye.ML.scalar_config import DEFAULT_KEY
 from InnerEye.ML.utils.metrics_util import binary_classification_accuracy, mean_absolute_error, \
     mean_squared_error, r2_score
 
 FloatOrInt = Union[float, int]
 T = TypeVar('T', np.ndarray, float)
 MetricsPerExecutionModeAndEpoch = Dict[ModelExecutionMode, Dict[Union[int, str], 'ScalarMetricsDict']]
+
+DEFAULT_KEY = "Default"
 
 
 def average_metric_values(values: List[float], skip_nan_when_averaging: bool) -> float: