+++ Tests/ML/test_metrics.py
@@ -18,7 +18,6 @@ from InnerEye.ML import metrics
 from InnerEye.ML.configs.classification.DummyClassification import DummyClassification
 from InnerEye.ML.configs.regression.DummyRegression import DummyRegression
 from InnerEye.ML.lightning_metrics import AverageWithoutNan, MetricForMultipleStructures, ScalarMetricsBase
-from InnerEye.ML.lightning_models import ScalarLightning
 from InnerEye.ML.metrics_dict import MetricsDict, get_column_name_for_logging
 
 
@@ -164,8 +163,8 @@ def test_get_column_name_for_logging() -> None:
 
 
 def test_classification_metrics() -> None:
-    classification_module = ScalarLightning(DummyClassification())
-    metrics = classification_module._get_metrics_computers()
+    config = DummyClassification()
+    metrics = config._get_metrics_computers()
     logits = [torch.tensor([2.1972, 1.3863, 0.4055]), torch.tensor([-0.8473, 2.1972, -0.4055])]
     posteriors = [torch.sigmoid(logit) for logit in logits]
     labels = [torch.tensor([1, 1, 0]), torch.tensor([0, 0, 0])]
@@ -203,8 +202,8 @@ def test_classification_metrics() -> None:
 
 
 def test_regression_metrics() -> None:
-    regression_module = ScalarLightning(DummyRegression())
-    metrics = regression_module._get_metrics_computers()
+    config = DummyRegression()
+    metrics = config._get_metrics_computers()
     outputs = [torch.tensor([1., 2., 1.]), torch.tensor([4., 0., 2.])]
     labels = [torch.tensor([1., 1., 0.]), torch.tensor([2., 0., 2.])]
     for output, label in zip(outputs, labels):